Add support for api version 6.1

+++ src/RxTelegram.Bot/Interface/BaseTypes/FileBase.cs
@@ -21,6 +21,6 @@
         /// <summary>
         ///     Optional. File size, if known
         /// </summary>
-        public int? FileSize { get; set; }
+        public long? FileSize { get; set; }
     }
 }
